chore: config: split globals in their own object
we were abusing objects to hold global settings.
It's time to start cleaning up, so we move globals in
their own object, and start adding p;roper JSON annotation.

internal-only change.
